+++ hypergeom.c
@@ -8,17 +8,17 @@
 #include "hypergeom.h"
 
 
-int terms_2F3(const struct coeffs_2f3, const double x, long double terms[]);
-int terms_1F2(const double coeff, const struct coeffs_1f2, const double x, long double terms[], const int start);
+int terms_2F3(const struct coeffs_2f3, const double x, double terms[]);
+int terms_1F2(const double coeff, const struct coeffs_1f2, const double x, double terms[], const int start);
 
 
-long double hyp1F2(const double a1, const double b1, const double b2, const double x)
+double hyp1F2(const double a1, const double b1, const double b2, const double x)
 {
         int k;
 
 
-        long double result = 0;
-        long double term = 1;                // Stores the running value of each term in the summation. From the definition of the Pochhammer symbols the value of the k = 0 terms is ONE
+        double result = 0;
+        double term = 1;                // Stores the running value of each term in the summation. From the definition of the Pochhammer symbols the value of the k = 0 terms is ONE
 
         // When (the absolute value of) 'term' becomes less than TOLERANCE the sum stops changing rapidly and we truncate it
         for (k = 0; (k < MAX_TERMS) & (fabs(term) > TOLERANCE); ++k)
@@ -33,12 +33,12 @@ long double hyp1F2(const double a1, const double b1, const double b2, const doub
 }
 
 
-long double hyp2F3(const double a1, const double a2, const double b1, const double b2, const double b3, const double x)
+double hyp2F3(const double a1, const double a2, const double b1, const double b2, const double b3, const double x)
 {
         int k;
 
-        long double result = 0;
-        long double term = 1;
+        double result = 0;
+        double term = 1;
 
         for (k = 0; (k < MAX_TERMS) & (fabs(term) > TOLERANCE); ++k)
         {
@@ -57,8 +57,8 @@ long double hyp2F3(const double a1, const double a2, const double b1, const doub
 int compare_terms(const void *pa, const void *pb)
 {
         // Use the void * pointers to get the long double values and take their absolute value.
-        long double a = fabsl(* (long double *) pa);
-        long double b = fabsl(* (long double *) pb);
+        double a = fabsl(* (double *) pa);
+        double b = fabsl(* (double *) pb);
 
         // The return values are chosen to give us descending order
         if (a < b)
@@ -71,19 +71,19 @@ int compare_terms(const void *pa, const void *pb)
 }
 
 
-long double together_hyp(const double coeff, const struct coeffs_1f2 c_1f2, const struct coeffs_2f3 c_2f3, const double x)
+double together_hyp(const double coeff, const struct coeffs_1f2 c_1f2, const struct coeffs_2f3 c_2f3, const double x)
 {
         int k;
 
-        long double terms[2 * MAX_TERMS];
+        double terms[2 * MAX_TERMS];
 
         int size = terms_2F3(c_2f3, x, terms);
         size = terms_1F2(coeff, c_1f2, x, terms, size);
 
         // Sort the terms in the array to reduce truncation errors
-        qsort(terms, size, sizeof(long double), compare_terms);
+        qsort(terms, size, sizeof(double), compare_terms);
 
-        long double result = 0;
+        double result = 0;
 
         for (k = 0; k < size; ++k)
                 result += terms[k];
@@ -95,11 +95,11 @@ long double together_hyp(const double coeff, const struct coeffs_1f2 c_1f2, cons
 /*
  * Calculate all of the terms of -2F3 and store them in the specified array
  */
-int terms_2F3(const struct coeffs_2f3 c, const double x, long double terms[])
+int terms_2F3(const struct coeffs_2f3 c, const double x, double terms[])
 {
         int k;
 
-        long double term = -1;          // - 2F3 is to be added to the final answer so each term will be negative
+        double term = -1;          // - 2F3 is to be added to the final answer so each term will be negative
 
         for (k = 0; ((k < MAX_TERMS) & (fabs(term) > TOLERANCE)); ++k)
         {
@@ -115,11 +115,11 @@ int terms_2F3(const struct coeffs_2f3 c, const double x, long double terms[])
  * The same array is passed which has already been populated by terms_2F3.
  * The int 'start' tells you where to start adding terms after the earlier population.
  */
-int terms_1F2(const double coeff, const struct coeffs_1f2 c, const double x, long double terms[], const int start)
+int terms_1F2(const double coeff, const struct coeffs_1f2 c, const double x, double terms[], const int start)
 {
         int k;
 
-        long double term = coeff;
+        double term = coeff;
 
         for (k = 0; ((k < MAX_TERMS) & (fabs(term) > TOLERANCE)); ++k)
         {
